requirements.txt
beautifulsoup4==4.11.2
matplotlib==3.4.3
nltk==3.3
numpy==1.22.4
python-decouple==3.7
requests==2.26.0
# spacy==3.1.4
Unidecode==1.3.6
spacy>=3.0.0,<4.0.0
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.5.0/en_core_web_md-3.5.0-py3-none-any.whl
